Update IPA info in troubleshooting.rst
The details on IPA image options were not previously detailed in
the troubleshooting documentation, and given that we have switched
from CoreOS to TinyIPA for testing, it is relevent to suppliment the
troubleshooting information. Updated formatting and placement in
an attempt to provide better clarity to the reader or user attempting
to remedy an issue.

Change-Id: Ic2f4acd8c4824ddc4c13dabec7d6b8cfb413b35b