Test timeout extension for inspection
The default introspection time for shade may be too short for some
environments. As such, extending the timeout to be longer to provide
nodes an opportunity to boot, check-in, and reply with results.

The library default, which I believe is 900 seconds, was also based
on basic tests where I determined that _most_ systems should check-in
with-in based on:

 - 300 seconds to POST
 - 300 seconds to PXE
 - 300 seconds to Boot and start IPA.

Change-Id: If17c9d7c6acc19ec05a33fa57cb4bbbef1fa43d0