Update Ansible version to stable-2.1
Updated the version to be installed to stable-2.1 as well
as updated some of the playbooks that were incorrectly
passing string arguments instead of dictionary arguments
for authentication since Ansible 2.1 has checks that
detect this.

Additionally, updated the task labels used for unsetting
authentiction parameters to be more consistent with the
state of the code.

Change-Id: I8c50cd2a722f2c8fee91fb6ef0761969dd3da138