Serialize localhost actions during final host deployment
Several tasks in the final playbooks run locally and there is a chance
to encounter race conditions when multiple tasks run in parallel and
all of them access or modify the same local files. We fix this by
splitting the final steps into multiple plays and serializing those
that run locally. Finally, the 'serial' keyword can only be used for
an entire play and not just for a subset of its tasks provided by a
role so previously that keyword did not have any effect.

Change-Id: Ib0a55dcd4b02e070a6be72207f0b99e8138e0e12