bifrost: inventory: Downgrade log level to 'debug' for csv parsing
Falling back to csv parsing is an intended behavior so downgrade the
log level to 'debug'. This also prevents the inventory script to print
the fallback message to stderr which can cause ansible-playbook to
fail to parse the rest of the output.

Change-Id: Ie79b70c9560ca627bc681948c9db3323ae9f447e
(cherry picked from commit 21f27311265fc43710f1c9793dd31fd38e9a64f6)