Provide default to uwsgi_install_prefix
due to how Ansible works, the expressions in `with_items` are still
evaluated even when run under `when` statement.

Due to this, the bifrost-keystone-install role failed in
tasks/bootstrap.yml when trying to place uwsgi services when keystone
was not actually installed as it was trying to evaluate
uwsgi_install_prefix.stdout while the task registering this var was
actually skipped.

This patch provides default to this evalation in the form of empty
string.

Change-Id: Id16f680853e5710deed5954c891b7e3ec83f16d2