Fix race in virbr0 startup
A race can exist with the virbr0 startup where libvirt eventually
starts virbr0, but this conflicts with our expliit attempt to start
it. As such, we shoulld only attempt to start it if it makes sense
to, and tolerate failures when it appears that libvirt started the
virbr0 interface while we were not looking.

Change-Id: I804102480a0e3df6cfa97ede1213404a7db1d0fe