roles: bifrost-ironic-install: Always download the latest iPXE files
Set the 'force' option in the get_url module to 'yes' to always download
the latest files if necessary. This aims to resolve iPXE problems
related to http://ipxe.org/err/040ee1. One of the solutions suggested
there is to use the latest version of iPXE. However, up to know, we
didn't update the files after the initial download which means that such
problems could persist indefinitely. iPXE files are really small, so
the download overhead is really minimal.

Change-Id: Ib8f7e4d74a33eeb4147732f7e161ba03de434e7b