Remove my_ip_address from use
With recent gate failures, the possibility exists that the
fact, via being a default, for my_ip_address, is being fullfilled
as part of template expansion as opposed to runtime.

Since it is possible that things are still in flight with a test
run, we should utilize the entire variable so it is expanded
when it is needed, after a refresh has occured.

Also moved the collection of facts to later after the completion
of package installation, into the bootstrap step since it is only
required in the bootstrap step, and not until users and database
access configuration has already been asserted.

Change-Id: I74f936ee0440d4afaedafc8eff5936196186d743