Remove deprecated test playbooks
The dynamic and dhcp test playbooks were supersceded in the last
cycle for the main playbook which contains all of the logic to
support the various cases and pass the appropriate configuration
for each case.

As an appropriate period of time has passed, it is time to remove
the deprecated test playbooks.

Change-Id: I2a7b94a7309e1dcf9232707439dd21c31538fd84