Move database setup to bifrost
Move openstack_citest database setup from project-config to bifrost.
The code is conditional on running in OpenStack CI - developers can
set up the database with setting ZUUL_BRANCH. This is the same logic
as used elsewhere in bifrost testing.

OpenStack CI will not setup databases anymore in project-config,
this needs to be done by the projects themselves now.

Change-Id: I8485751eda745841c1436ecd766c9bb95446f407
Needed-By: I339d2a301a76ad34ea4bcbfb26c13a21ecd86b6f