Workaround for network start race condition
Add a workaround for the libvirt network start race condition bug. In
some cases the libvirt default network starts in an inactive state
with the bridge interface down. Trying to start the network via libvirt
fails because the network interface already exists. This condition is
checked for and ignored but causes the Bifrost deployment test to fail
later when templating ironic.conf as the configured network interface
has no IP address.

This change checks for the condition where the libvirt default network
is inactive and attempts to delete the associated bridge interface
before starting the libvirt network. This has been run through the CI
job and appears to resolve the issue.

Change-Id: Icb2baf0975ca3feb6041b2f2da9702ad6cc9fa06
Closes-Bug: #1660953
Related-Bug: #1650025