playbooks: roles: bifrost-ironic-install: Verify checksum for IPA files
Use the newly published checksum files to verify the integrity of the
downloaded IPA files. If verification fails, we try a few more times
before we give up since that would either mean that there is a
connectivity issue on our end or a problem with the uploaded files.
Checksum verification is enabled by default but it can be easily
disabled in case we do a non-upstream build with no checksum files
available.

Depends-On: I66afa3682ee1461ba1cdb68a134eaa4fd85af4e0
Change-Id: I7b3cad1b42e297b87ca839d66a16303250e27404