playbooks: roles: bifrost-ironic-install: Retry failed PXE downloads
Downloading PXE files could fail for various reasons which may not
always be valid so try to workaround that by simply trying to get the
files a few more times before giving up. This aims to workaround the
following issue which appears from time to time in Jenkins builds
even though the remote certificate and the host's software remain the
same.

"Failed to validate the SSL certificate for boot.ipxe.org:443. Make sure
your managed systems have a valid CA certificate installed. You can use
validate_certs=False if you do not need to confirm the servers identity
but this is unsafe and not recommended. Paths checked for this platform:
/etc/ssl/certs, /etc/pki/ca-trust/extracted/pem, /etc/pki/tls/certs,
/usr/share/ca-certificates/cacert.org, /etc/ansible"

However, it may also assist with general networking issues as well.

Change-Id: Ia2351ed2c7f1dfe26c17770250d150f26432086b