Ironic inventory source duplicates MACs
When using the 'ironic' Bifrost inventory source to obtain an Ansible
inventory from ironic, if any nodes have multiple ports then the
resulting inventory contains a 'nics' list variable where each item has
the same MAC address.

This change fixes this issue by reinitialising each new_nic variable
such that all items in the nics list do not reference the same dict.
We also introduce a couple of unit tests for the ironic inventory
source - one to test the simple case of a single node with a single NIC,
another to test the broken case of a single node with multiple NICS.

Change-Id: I173fbbdbebafc05536a153dcf59e4b2f37aebdbc
Closes-Bug: #1671014