Resolve evaluation of bifrost-keystone-client-config
The owner parameter for some of the tasks for the Keystone
client configuration does not have a default fallback, which
results in a failure if SUDO_USER is not set in the environment.

This patch adds the missing default fallbacks so that it can safely
evaluate to root if it missing from the environment.

Related-Bug: #1674483

Change-Id: I31a2a33e2fb3b004dfec179426c7d1b4c72301d2