Fix endpoint creational in ironic
The openstack command, when an endpoint is checked directly
returns a code of 1 if no endpoint is found. This by default
for ansible is a failure, and causes the task to fail, however
we check for the return code later. We should ignore the errors
upon these information gathering tasks being executed.

Change-Id: I4cb7eb2878f1495a17821b895765093205f43c98