Support storage of introspection data in Nginx
As an operator I want to be able to persist raw and processed
introspection data so that I am able to view it at a later date.

As an operator I want to be able to persist raw and processed
introspection data so that I am able to reprocess the data after the
initial inspection process has completed.

In the absence of swift, we can use the bifrost nginx web server -
masquerading as an object store - to store raw and processed
introspection data for nodes.

This allows introspection data to be retrieved and reprocessed after the
initial inspection has completed. This can be useful when the processing
pipeline or introspection rules are changed.

Change-Id: Ia2bd16080594e854054f380d4f7670eaea98e82b
Closes-Bug: #1685879