Allow log directories to be configured
In some cases it is useful to be able to configure the directory into
which log files are written by bifrost services.

Of particular note is the kolla-ansible project which typically stores
log files in a docker volume mapped to /var/log/kolla/ inside
containers. Currently the bifrost_deploy container is unable to persist
its logs for this reason.

This change enables configuration of the log directory path for the
following services:

- ironic-api
- ironic-conductor
- ironic-inspector
- nginx

Change-Id: I452052fa2f9a55fac2f00b454a5ac70464c9d981
Closes-Bug: #1685871