Revert "Remove useless paramiko requirements"
This reverts commit 8e226b826f35826dad6c673d93991f459de51c33
which removed paramiko from requirements.txt.

However, paramiko is still required on a system level for ironic.

What occurs is Ansible's paramiko gets installed. When ironic is
installed by that user, pip sees the user's specific paramiko
install, and assumes that in part of the dependency calculation.

This fails when ironic is installed to the ironic user. As such,
the cleanest approach, for now, is to ensure that paramiko is
installed system wide.

Change-Id: Ifcb2715864e8d567aedcb3126afa23d654532e63
Closes-Bug: #1695996