Add bindep.txt file
OpenStack CI uses bindep to figure out which packages need to be
installed on the host. However, when a bindep.txt file is not
available, a default one is used which may or may not be good enough
for all cases. It's best to have control over the packages which are
going to be installed on the host and as such we add our own bindep.txt
file.

Change-Id: Idab36eb82d916183f2fb40d326d095c5a8ee34b2