bifrost-create-vm-nodes: Do not use KVM VMs by default
The default option is to not use KVM for virtual machines but CentOS is
overriding that and uses KVM as the default option. This leads to
inconsistencies and causes some issues when testing bifrost on
hypervisors which do not support nested virtualization (ie virtualbox)
so use the qemu-system-x86_64 emulator instead which does not use KVM
but dynamic binary translation instead.

Change-Id: I18dab3e74d1395a6c7d129701907424efaa5f735