Split inspection conditionals depending on auth mode
There is a bug in the inspection module, which needs
to be corrected, although that is going to take potentially
a substantial amount of time given ansible's release cycle.

In the mean time, we're just going to have to work around it
until we're able to get the fix landed in ansible.

Change-Id: I70fc4b54c6cd815bdb755819ad28215714c71260