Correct rabbitmq port usage
Several messages where we had a hard coded
wait for a hard coded port, for rabbitmq,
were present.

Replaced all but one of the instances with
variable usage, the remainder I added a TODO
as more refactoring of that file is required
and it should be done at the same time.

Change-Id: I5019424c79763254854fdaace3da2753a05cd94f