Different pip install with/without venv
When testing the scripts with newer ansible version, it was failing
to install packages with pip and without venv. The reason is because
even the venv vars are set to omit, the module is still acting as it
had an empty virtualenv, causing this to fail. So just create a
different step, without the vars being set.

Change-Id: I0b268a170e69ccc36f90d875f25453d3874b22b0