Unquote empty braces inside jinja blocks
Quoting empty braces inside jinja blocks (i.e. {{ '{}' }}) tells jinja to treat
the braces as a string instead of as an empty dict. Since these occur in jinja
blocks, the yaml processor has already interpretted them as strings, and there is
no threat of the yaml processor interpretting the braces as a yaml object.

As-is, the quoted strings cause errors such as:
  FAILED! => {"failed": true, "msg": "|combine expects dictionaries, got '{}'"}

Change-Id: I4f859d8e3b9eff7ce962d7690c1ceea3e6229fe5