Correct minor pike branch config issues
Bifrost has many knobs, and we've not yet made a list
of things that need to be followed-up with each stable
release.

As such, we need to ensure we install the appropriate
stable branch versions from source.

Set the branch, set the IPA image to be the pike image
and change inspector to install from source, as the default
master branch behavior is to use the released version for
testing.

Change-Id: Ia5b63602949db9c11cce3ae40f753b01d59084cc