Make tinyipa and tinyipa-opensuse-423 work for other repos
This is a VERY tiny change to just make these two jobs work when
triggered from the shade repo. The larger fix is to rework all of the
integration jobs to use the new devstack base job ... but wow that'll be
more work than I feel lke doing on a Sunday morning.

Change-Id: Ic192bd6b5fa16726d6c03e83e752bc8beee81720