Install shade after ironic
Due to a upper constraint, the installation of ironic is
resulting in keystoneauth1 being downgraded. Since bifrost
is about using shade from the current state of master for
compatability reasons, we need to change the install order.

Also, in the course of discovering this, fixing what appears
to be a bad conditional statement where the when was missing
which resulted in an error when testing with a newer
keystoneauth1 library.

Change-Id: I81a0261093cacb636a5c958c32ca31bd7d4c0042