roles: bifrost-create-vm-nodes: Set NIC model to 'virtio'
The role already uses the virtio driver to improve the disk peformance.
We can also use the virtio driver for the NIC device in order to improve
the network performance of the test VM. However, users may want to
change that so we are adding a new test_vm_nic Ansible variable for
that.

Change-Id: I9e5c9a33f1f6903d39d0fa2721288d084071e733