Allow missing ironic policy.json sample
Since ironic change I10a4987204eb335f1fed29d4b1d7f3e075aa1792, the
sample policy.json is not present in the ironic repository. Since policy
is now in code, it is not required. This change only copies the sample
file to /etc/ironic when it exists in the ironic repository.

Change-Id: Iecdbd90b4c94556f5d0803e4340dfd6b473792c7