Enabled hardware type 'ucs' does not exist
During classic driver deprecation [1], the ucs hardware type was added
to the list of default enabled hardware types. This hardware type does
not exist - the Cisco UCS hardware types are cisco-ucs-managed and
cisco-ucs-standalone. This causes conductor to fail to start.

This was not caught during testing as only the ipmi ironic hardware type
is configured when testing. Now that VBMC is in use, it is no
longer necessary to use a different set of drivers for testing.

This change enables the cisco-ucs-managed hardware type by default, and
uses the default set of drivers and hardware types for testing. It also
adds configuration options for setting the enabled boot, power and
management interfaces, with default values including interfaces required
by the ilo and ucs hardware types.

[1]
https://github.com/openstack/bifrost/commit/46453661badf83c2c45a008908d4476de6785ccd

Change-Id: Icfa89a2555966c6fbfd06186fdd17d852bc5ae7a
Closes-Bug: #1752975