Install openstack client, inspector in clouds.yml
Installs the openstack client unconditionally, not just when keystone is
enabled. This allows us to communicate with ironic inspector, and a new
bifrost-inspector entry has been added to clouds.yml in the noauth case
to support this.

Also adds a play to test-bifrost.yaml that exercises the openstack
client by performing a node list and querying inspector rules.

Change-Id: I19dfe3f63ba2c7fea7ce5881d7c8d0dff8f11264