Use endpoint rather than url in clouds.yaml with no keystone
When keystone is not in use, the clouds.yaml file generated by bifrost
uses an auth_type of "none". This does not appear to work, and results
in the following error (even with OS_CLOUD=bifrost):

Missing parameter(s):
Set a cloud-name with --os-cloud or OS_CLOUD

This change modifies the "url" parameter to be "endpoint", which works
correctly.

Change-Id: I9b84739dcb90b6482b332676244c68642042dfd4
Closes-Bug: #1754070