Fix bifrost CI
Bifrost's CI is broken by two items:

Ansible 2.5's pause module seems to be
more syntax strict, and we might have been
missing the pause logic all along due to the
when being down a level too far in the yaml.

Additionally, fedora core 26 support in CI
is no longer available, and as such disabling
for the time being.

This should fix CI.

Change-Id: I37df36d5f1bf13c833feda9e0a7f0a288a18a097