Only install libvirt-python and python-lxml via pip
These are required by the ansible net_virt module, and also virtualbmc.
As such, they will be upgraded when virtualbmc is installed. Pip 10 will
not allow this, if are already installed by the system package manager.
Let's only ever install them via pip as that's the version we want in
the end anyway.

Change-Id: Ia209f22170905a5acd23eeb82cbdec9a537bf80b