Only install libvirt-python and python-lxml via pip
These are required by the ansible net_virt module, and also virtualbmc.
As such, they will be upgraded when virtualbmc is installed. Pip 10 will
not allow this, if are already installed by the system package manager.
Let's only ever install them via pip as that's the version we want in
the end anyway.

Additionally correct a typo in the test playbook that was preventing
changes on queens to land.

Change-Id: Ia209f22170905a5acd23eeb82cbdec9a537bf80b
(cherry picked from commit 50c7edb9d8c5383303207026d30b4badc0c5f398)