Remove legacy-fedora-26 job
We'd like to remove this nodeset from openstack-zuul-jobs, but because
this configuration is still referenced by bitfrost we cannot.

For now, we can just delete this job, since it has been commented out,
and can be reverted once somebody has the time to update it to use the
latest nodeset of fedora.

Change-Id: I3ace0f8d3aaf2b0e27349415f1571e4811b6ccf1
Signed-off-by: Paul Belanger <pabelanger@redhat.com>