playbooks: roles: Drop explicit 'delegate_to' options
Bifrost instructions suggest to use 'connection: local' when calling any
of the roles. However, there are scenarios which the roles can be
delegated to some other host. The explicit 'delegate_to' statements
prevent us from doing that and they are not necessary if we use
'connection: local' anyway. As such we drop the explicit statements to
make the roles a bit more flexible.

Change-Id: Ifc983992125d203899d74b8bf997f9f58f528978