playbooks: roles: Do not force dependency upgrades
pip >= 10 is very strict when it comes to satisfying dependency
requirements. If python-foo is installed from the distro repos but
a higher version is available in pypi, then pip will try to upgrade the
package and fail with the following error:

 It is a distutils installed project and thus we cannot accurately determine
 which files belong to it which would lead to only a partial uninstall.

As such, lets make pip less aggressive by allowing existing packages to
satisfy dependency requirements.

Change-Id: I3a28abd4a68676fda180694a030ea40c7552d8d2