Remove pre-emptive boot.ipxe copy
Early on in bifrost, we used a static template for
a short time. As time evolved, so did the template,
and we were able to rely upon the conductor placing
the boot.ipxe file upon startup.

Ironic since performed the copy upon driver init
as opposed to once a deployment has been requested.

As it is no longer needed, we shouldn't put an invalid
ipxe boot configuration in place. Removing legacy copy.

Change-Id: I7f2fc2afca6c8ee0f18efe8aa461bf906794e049