playbooks: Convert venv_python_path to proper role
The venv_python_path tasks were in the roles directroy but we were
including it as a normal file with a list of tasks so it was not really
an Ansible role. This does not work well when we use the Bifrost roles
outside of bifrost since Ansible does not know where to find this file
for inclusion. Our best chance is to convert it to a simple role which
can be statically imported.

Change-Id: I82fdfe48f9f2eb631165ab630157bc378ddc631d