Perform ironic online data migrations
Online data migrations are necessary for performing an upgrade of
ironic. Without this change, an error such as the following may be seen
in task 'bifrost-ironic-install : Upgrade ironic DB Schema'.

The database is not compatible with this release of ironic (10.1.7).
Please run "ironic-dbsync online_data_migrations" using the previous
release.

Change-Id: I685efd14bf3567a126311b676a50b0abb0f043db
Story: 2004670
Task: 28658