Use a virtualenv in openSUSE CI testing
Currently all of the opensuse 150 jobs are failing due to being unable
to install virtualbmc via pip. The failure signature is:

Cannot uninstall 'six'. It is a distutils installed project and thus we
cannot accurately determine which files belong to it which would lead to
only a partial uninstall.

The usual solution to this problem is to install Python packages in a
virtual environment. This change modifies the zuul CI jobs to use a
virtual environment for all openSUSE jobs.

Change-Id: I2840ec32cffcaf5d7f32a91907d51c1a500f19f1