Allow Ironic URL to use env vars in bifrost_inventory.py
Currently the URL used in bifrost_inventory.py is hard-coded to
http://localhost:6385, which might not be valid if we are accessing a remote
API or the service is bound to a specific IP.

The URL can be specified using either OS_URL or IRONIC_URL, with the former
taking precedence.

This patch also adds support for using clouds.yaml with bifrost_inventory.py.

Change-Id: I8741feb05a0a1cdd55f2e1f8d4cea8fcfefed192