Pin proliantutils version
This avoids an issue where the version installed could be incompatible
with ironic, or other installed libraries.

In particular, this was seen with proliantutils 2.8.2, which is
incompatible with the version of sushy-lib on the stable/rocky branch.

Change-Id: I02f1e247ca2333b2e90586e7311395e7bd94938c
Story: 2005310
Task: 30214
(cherry picked from commit 348b45aca8dc850840b3a3b656a6a8fc8f99bf47)