Migrate to Zuul v3 native jobs
This corrects the path handling in the CI and removes a lot of
redundancy in the CI definitions. Also adds fixed log collection
from https://review.opendev.org/#/c/649526/ to make the CI pass.

Removes handling of ZUUL_CHANGES. It is not present in new jobs,
and its use case in Bifrost is completely unclear.

Eventually, we should rewrite the whole test-bifrost.sh in ansible,
but that's a much bigger task.

Change-Id: Id893e816a7806e7d75282628817288d8a34ab8af
(cherry picked from commit 4bb71fc481eb66463084cf07bdd781c98b67100b)