Disable bifrost jobs so CI can be fixed
tl;dr: When ironic dropped py2 support, bifrost's
CI reached a point where it could not be fixed without
a massive changeset being created with all required
actions to purge all traces of python2.

In order to work through this, and hopefully make things
better, this change disables the bulk of the CI jobs, marks
the existing primary jobs we want to succeed to be non-voting,
and creates a bionic job as we should ideally already be on
bionic.

The overall plan being to iterate through the jobs and get them
working one at a time as opposed to try and make them all
functional in one changeset.

Change-Id: I9a1e29d1c8bcc4f3cf00df81405f82c22fd10188