Make a copy of requirements for local testers
The requirements repository is provided to us in most CI
testing scenarios, however we also want to force a constraint
upon installing python libraries for test VM creation dependencies.

As such lets at least get a copy of requirements up front when
testing with test-bifrost.sh or when using vagrant.

Change-Id: Ic6a5b8f7cdbf6f4c6233848bab1bfd71acc9d973