Skip status code check and make vagrant install quicker
Under some failure conditions, the status code is not returned.

Also skip the agent/os build steps for install inside of vagrant
which can add a substantial amount of time to the build process.

And minor revision to the test-bifrost.yaml conditional as the
pattern was traditionally "is defined", but when running the test
scripts, most things are defined, so it was just a side effect of
changes being purely tested in CI over time.

Change-Id: Icf132b71b243f9fce9d1c2afca656299befcfabc