Use upper-constraints for ironicclient, sushy and DIB
The goal of upper-constraints is to avoid breakages with versions
known not to work. Lack of upper-constraints for ironicclient is
currently breaking stable branches by pulling packages that are
not Python 2 compatible.

If we need recent versions of ironicclient, we should use source
installation instead.

To make it work with packages that are constrained themselves,
this patch splits the source-based installation into two steps:
installing requirements (constrained) and installing the project
itself (unconstrained).

Change-Id: Ie5f62655248c40885fdcd6fa87180a6952f145fc