Fix test for python3-venv pkg install
The test to check if the python3-venv package is needed is wrong
and it will always fail, asking to install the package in systems
where that doesn't exist, like CentOS or Fedora.
The python3-venv package is always needed for ubuntu bionic, so
installing it all the time if we're using that distro.

Change-Id: I6fe765609cc0561f359986f488268e79cca469e2