Run vbmcd as a Systemd service
Virtualbmc 2.0 dropped the daemon autostart feature. This means we need
to explicitly start it. Running it as a Systemd unit, code adapted from
Tenks.

In order to allow Systemd to execute the daemon on systems with SELinux
enabled, it must be installed to the system rather than a virtualenv.

Change-Id: I66e8e5f5a9c16ac90e132074c826d508458ae4d1
(cherry picked from commit 6513e5a6dc92dff06e7d5caeadf14db434b3648f)