Ensure DIB is installed from source
Bifrost clones the DIB repo and installs it. However, after running
install.yaml pip freeze shows diskimage-builder 2.33 is installed.

This is caused by the recent addition of IPA builder, which has its own
dependency on diskimage-builder.

This change fixes the issue by swapping the order of installation of IPA
builder and DIB, to ensure DIB is installed last.

Change-Id: Ifd67adaa98e920f42c80d239349ae61d7d610f66
Story: 2007401
Task: 38999