Remove --ignore-installed from CentOS 8 pip 19 install
In the kolla bifrost container image this results in a broken pip
installation. The pip 19 package is downloaded, but this message is
printed at the end:

Successfully installed pip-20.0.2

The subsequent 'pip freeze' gets the following error:

ImportError: cannot import name 'PackageFinder'

This change removes the --ignore-installed argument, which results in
the virtualenv getting pip 19 as expected.

Change-Id: Idf2b03a4772111f63f79203f63ea95f805e7a43d