Correct handling enable_venv in pip_install.yml
The way we pass it, it is a string, not a boolean. While ansible
is handling this situation for now, it's deprecated and should
not be relied on. Convert enable_venv to boolean explicitly.

Remove a redundant fact setting and redundant default() from
enable_venv (it is set in role defaults).

Also synchronize pip_install.yml to bifrost-keystone-install.

Change-Id: I35be662c4a60b4085b22d7e01b1691597f82c732
(cherry picked from commit bb807b2288551bae02dab4f34da29226b3eba5b2)