CI: reduce memory and disk for testing nodes
There is not reason to use 10 or 12G of disk for Cirros.

Also change the RAM to 512M, although we'll have to revert it back
on master and ussuri once we switch to DIB images by default.

Conflicts:
	scripts/test-bifrost.sh

Change-Id: Ibada78686ffb4db1d8dae05acf980081d5265fc6
(cherry picked from commit 85e1b8207885d64daf986158dba8768ae49b257e)