Use firewalld to open ports on CentOS and RHEL.
The current code does not properly open ports in the libvirt zone
and firewalld is the default in CentOS/RHEL 8.

We will update Fedora once we switch to a newer version that
defaults to firewalld (30 apparently does not).

Change-Id: If6b3d13e2b4bf0e71b0968850e05ea93efd6f189
(cherry picked from commit deaa4d7186c9937717424aafa1b3fef6af96fcf6)