Removing libselinux-python package from Fedora dependencies
In the past releasaes of Fedora, bifrost setup had a dependency on
libselinux-python package. In Fedora 32 this package no longer exists
due to transition to Python3 and the absence of this package in
Fedora 32 repos this causes the playbook to fail.

This change removes libselinux-python from Bifrost Fedora dependencies
and resolves the issue.

Change-Id: Id767e76bb0afb91f4176fd02ad1fa33f17325df2