Create symlink to ansible collections on install
Bifrost requires Ansible Collections for certain tasks including
enrollment of nodes. In order for bifrost playbooks to be able to
utilise Ansible Collections, a symlink or an environment variable
needs to point to their install path. This change adds creation of
the symlink to the setup script and a corresponding documentation
update.

Change-Id: I07804a88ed11b0adb6c398743905725c6cdb90d4